pyproject.toml
[tool.poetry]
name = "pod-bot"
version = "0.1.0"
description = "Automated podcast clipper bot"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
yt-dlp = "2024.7.23"
openai-whisper = "20231117"
openai = "1.12.0"
ffmpeg-python = "0.2.0"
pillow = "10.1.0"
python-telegram-bot = "20.6"
flask = "3.0.0"
pydantic = "2.5.0"
boto3 = "1.34.0"
pandas = "2.1.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.11.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
